Refactored functions into separate files
